urdu-ocr-backend/requirements.txt
Flask
google-cloud-vision
google-cloud-storage
google-cloud-firestore